Made primes example even MORE efficient & fixed build
I looked back on my contribution to this project,
and realized I had made a mistake! `n / base 10 length of n`
is not the best way to overestimate primes, `n / base 3 length of n`
is! So i fixed that, and it works :-)

I also fixed the travis CI build failing, because that was annoying